refactor: share item and effect context builders across actor sheets
アイテム一覧と効果の区分けはアクター種別に依らない形なので、
context/character.ts から context/actor.ts へ出す。ハウリング由来の効果を
除くかは引数にする — 専用区分を持つ共鳴者だけが除き、持たないシートでは
乗ってしまった効果も隠さず出す。
